Finish cleanup of obsolete tasks.common
